Refactor train into tuned model classmethod
